requirements.txt
fastapi
uvicorn
pydantic-ai
chromadb
google-generativeai
python-dotenv
pydantic-settings
